requirements.txt
django 
requests
django-shopping-cart